7) Used heatmaps to improve the evaluation function
Instead of valuing each peace naively, the code now changes the value
of each piece depending on which square it's on. So a double look up is
now used to get the value of a piece.

I also added a message at the end to output the result of the game, and
I renamed some variables to ensure the code is as readable as possible.