8) Better minimax
Using a mini, maxi and general evaluation function allowed for a faster
and more clearer set of functions.

I also replaced the checkmate checks with outcome checks as that was an
error I spotted in testing (draws worked badly). The evaluation also
gives a bonus to the player whose turn it is and checkmates quicker are
evaluated higher by the attacking and lower by the defending side.