11) Created the quiesence search
By allowing nodes to ensure they aren't evaluating at a critical stage
in the game they run a quiesence search that makes sure only 'quiet'
positions are evaluated.

This also required the making of a function to check the value of
captures without going through the whole board to use as an orderer.

I also cleaned up code, comments and variable names.